requirements/local.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==4.1.0
black==23.3.0
cachetools==5.3.1
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-coreapi==2.3.0
django-cors-headers==4.0.0
django-cron==0.6.0
django-debug-toolbar==4.1.0
django-extensions==3.2.1
django-filter==23.2
django-health-check==3.17.0
django-jazzmin==2.6.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.1
facebook-sdk==3.1.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.3.0
plotly==5.14.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tomli==2.0.1
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
zxcvbn==4.4.28
